Split tests to one file per class
